make a refactor of gen_proof
