+++ src/lib.rs
@@ -1,3 +1,3 @@
 pub mod merkle_tree;
 pub mod merkle_error;
-pub mod merkle_proof;
+pub mod proof_element;
